chore: add native wind and configure app
- deleted App.js file
- restructured the project layout
- added native wind package
- configured native wind
