+++ app/_layout.jsx
@@ -1,20 +1,12 @@
-import { StatusBar } from 'expo-status-bar';
+import { Slot, Stack } from 'expo-router';
 import { StyleSheet, Text, View } from 'react-native';
 
-export default function App() {
+export default function RootLayout() {
 	return (
-		<View style={styles.container}>
-			<Text>Default Layout</Text>
-			<StatusBar style="auto" />
-		</View>
-	);
+		<Stack>
+			<Stack.Screen name="index" options={{ headerShown: false }} />
+		</Stack>
+	)
 }
 
-const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: '#fff',
-		alignItems: 'center',
-		justifyContent: 'center',
-	},
-});
+
