+++ package.json
@@ -9,21 +9,23 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "expo": "~51.0.20",
-    "expo-status-bar": "~1.12.1",
-    "react": "18.2.0",
-    "react-native": "0.74.3",
-    "react-dom": "18.2.0",
-    "react-native-web": "~0.19.10",
     "@expo/metro-runtime": "~3.2.1",
+    "expo": "~51.0.20",
+    "expo-constants": "~16.0.2",
+    "expo-linking": "~6.3.1",
     "expo-router": "~3.5.18",
+    "expo-status-bar": "~1.12.1",
+    "nativewind": "^2.0.11",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-native": "0.74.3",
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
-    "expo-linking": "~6.3.1",
-    "expo-constants": "~16.0.2"
+    "react-native-web": "~0.19.10"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "tailwindcss": "^3.3.2"
   },
   "private": true
 }
